Moved log_fact() and log_choose() to where it is called more frequently
