+++ sm-spline.c
@@ -15,6 +15,36 @@ static omp_lock_t thread_lock;
 extern int spline_pts;
 double log_ad_step;
 
+double log_fact(int n) {
+  static int n_max = 0;
+  static double *lookup_table = NULL;
+  int i;
+
+  if (n < 0) return -DBL_MAX;
+
+  if (n == 0 || n == 1) return 0.;
+
+  if (n > n_max) {
+    RA(lookup_table, sizeof(double)*(n+1));
+    if (n_max == 0) {
+      lookup_table[0] = 0;
+      n_max = 1;
+    }
+    for(i=n_max;i<=n;i++)
+      lookup_table[i] = lookup_table[i-1] + log(i);
+    n_max = n;
+  }
+
+  return lookup_table[n];
+}
+
+double lchoose(int n, int k) {
+
+  if (n == 0 && k == 0) return 0.;
+  if (k > n || n == 0) return -DBL_MAX;
+  return log_fact(n) - log_fact(k) - log_fact(n-k);
+}
+
 double spline_interpolate(spline_t *spf, double x) {
   int i;
   double sp_y;
